Cast the format marker's var comparison
juce::var has several viable implicit conversions, so `var >= int` does not
resolve. Caught by the platform build; the standalone harness cannot see it
because it does not link JUCE.

+++ Source/PluginProcessor.cpp
@@ -253,7 +253,10 @@ namespace
 
     void migrateStateFormat (juce::ValueTree& tree)
     {
-        if (tree.getProperty (stateFormatProperty, 1) >= currentStateFormat)
+        // The cast is not decoration: getProperty returns a juce::var, and var against
+        // an int has several viable implicit conversions, so `>=` is ambiguous
+        // without it.
+        if (static_cast<int> (tree.getProperty (stateFormatProperty, 1)) >= currentStateFormat)
             return;
 
         if (auto mixChild = tree.getChildWithProperty ("id", "mix"); mixChild.isValid())
